+++ sea-orm-macros/src/derives/attributes.rs
@@ -30,5 +30,32 @@ pub mod field_attr {
         pub from: Option<syn::Lit>,
         pub to: Option<syn::Lit>,
         pub fk_name: Option<syn::Lit>,
+        pub condition_type: Option<syn::Lit>,
+    }
+}
+
+pub mod related_attr {
+    use bae::FromAttributes;
+
+    /// Operations for RelatedEntity enumeration
+    #[derive(Default, FromAttributes)]
+    pub struct SeaOrm {
+        ///
+        /// Allows to modify target entity
+        ///
+        /// Required on enumeration variants
+        ///
+        /// If used on enumeration attributes
+        /// it allows to specify different
+        /// Entity ident
+        pub entity: Option<syn::Lit>,
+        ///
+        /// Allows to specify RelationDef
+        ///
+        /// Optional
+        ///
+        /// If not supplied the generated code
+        /// will utilize `impl Related` trait
+        pub def: Option<syn::Lit>,
     }
 }