+++ sea-orm-migration/src/prelude.rs
@@ -9,5 +9,5 @@ pub use async_trait;
 pub use sea_orm::{
     self,
     sea_query::{self, *},
-    DbErr, DeriveMigrationName,
+    ConnectionTrait, DbErr, DeriveMigrationName,
 };