+++ tests/common/features/event_trigger.rs
@@ -1,6 +1,6 @@
 use sea_orm::entity::prelude::*;
 use sea_orm::{
-    sea_query::{ArrayType, ColumnType, RcOrArc, ValueType},
+    sea_query::{ArrayType, ColumnType, ValueType},
     TryGetError, TryGetable,
 };